[FE] REFACTOR: store/project, store/task 일부 수정
