+++ client/src/store/task.js
@@ -1,5 +1,6 @@
 import taskAPI from "../api/task";
 import { isToday } from "@/utils/date";
+import projectAPI from "../api/project";
 
 const state = {
   newTask: {},
@@ -11,9 +12,7 @@ const getters = {
   currentTask: (state) => state.currentTask,
   todayTasks: (state) => state.tasks.filter((task) => isToday(task.dueDate)),
   expiredTasks: (state) => state.tasks.filter((task) => !isToday(task.dueDate)),
-  taskCount: (state) => {
-    return state.tasks.reduce((acc, task) => acc + task.tasks.length, state.tasks.length);
-  },
+  taskCount: (state) => state.tasks.length,
 };
 
 const mutations = {
@@ -42,6 +41,29 @@ const actions = {
       commit("SET_ERROR_ALERT", err.response);
     }
   },
+  async addTask({ dispatch, commit }, task) {
+    try {
+      await taskAPI.createTask(task);
+      await dispatch("fetchCurrentProject", task.projectId);
+      await dispatch("fetchAllTasks");
+      await dispatch("fetchProjectInfos");
+      //commit("ADD_TASK_COUNT", task.projectId);
+    } catch (err) {
+      commit("SET_ERROR_ALERT", err.response);
+    }
+  },
+  async updateTaskToDone({ dispatch, commit }, { projectId, taskId }) {
+    try {
+      await taskAPI.updateTask(taskId, { isDone: true });
+      await dispatch("fetchCurrentProject", projectId);
+      await dispatch("fetchAllTasks");
+      await dispatch("fetchProjectInfos");
+
+      commit("SET_SUCCESS_ALERT", "작업을 완료했습니다.");
+    } catch (err) {
+      commit("SET_ERROR_ALERT", err.response);
+    }
+  },
   async updateTask({ commit, dispatch }, task) {
     try {
       taskAPI.updateTask(task);
@@ -54,6 +76,24 @@ const actions = {
   },
   startDragTask({ commit }, { task }) {
     commit("SET_DRAGGING_TASK", task);
+  },
+  async changeTaskPosition({ rootState, dispatch, commit }, { orderedTasks }) {
+    const { draggingTask, dropTargetSection } = rootState.dragAndDrop;
+
+    try {
+      await taskAPI.updateTask(draggingTask.id, {
+        sectionId: dropTargetSection.id,
+      });
+      await projectAPI.updateTaskPosition(dropTargetSection.projectId, dropTargetSection.id, {
+        orderedTasks,
+      });
+      await dispatch("fetchCurrentProject", dropTargetSection.projectId);
+      await dispatch("fetchAllTasks");
+
+      commit("SET_SUCCESS_ALERT", "작업 위치가 변경되었습니다.");
+    } catch (err) {
+      commit("SET_ERROR_ALERT", err.response);
+    }
   },
 };
